feat(T019): fix Android debug build — add Android SDK to nix flake, fix Kotlin type error
Add androidenv.composeAndroidPackages to flake.nix with platform 34 and
build-tools 34.0.0. Set ANDROID_HOME and accept unfree/license in flake
config so nix develop -c works without --impure. Fix Short/Int type
mismatch in YubikeyManager.handlePinError. Phase 7 checkpoint met.

+++ android/app/src/main/kotlin/com/agentrunner/yubikey/YubikeyManager.kt
@@ -192,7 +192,7 @@ open class YubikeyManager(context: Context) {
      * SW 0x6983 = PIN blocked.
      */
     private fun handlePinError(e: ApduException): Nothing {
-        val sw = e.sw
+        val sw = e.sw.toInt()
         if (sw == SW_PIN_BLOCKED.toInt()) {
             clearPin()
             throw PinBlockedException("PIN is blocked. Use ykman piv access unblock-pin to recover.")
